Cache user data in browser
The implementation embedded in this commit checks localStorage for any previous data whether for the grades or years so that it can be loaded initially.
In case there doesn't exist any data the default is loaded.
Reset buttons were added to the Heading of the two panels.
The collapse action of the GPA settings panel was modified so that it doesn't collapse in case of pressing reset button.

fixes #2